Split model and view files
